don't overwrite repo db's on init re-run

+++ init.py
@@ -56,8 +56,10 @@ def dummy_repo_db(repo_path):
 	repo_path = os.path.join(root, repo_path)
 	print os.path.join(repo_path, 'aurstaging.db')
 	try:
-		subprocess.call(['tar', 'czf', 'aurstaging.db.tar.gz', '--files-from', '/dev/null'], cwd=repo_path)
-		os.symlink('aurstaging.db.tar.gz', os.path.join(repo_path, 'aurstaging.db'))
+		if not os.path.exists(os.path.join(repo_path, 'aurstaging.db.tar.gz')):
+			subprocess.call(['tar', 'czf', 'aurstaging.db.tar.gz', '--files-from', '/dev/null'], cwd=repo_path)
+		if not os.path.exists(os.path.join(repo_path, 'aurstaging.db')):
+			os.symlink('aurstaging.db.tar.gz', os.path.join(repo_path, 'aurstaging.db'))
 	except OSError:
 		# File exists
 		pass
